register nil check
